+++ src/main/java/org/anstis/client/grid/widget/renderers/mergable/MergableGridRenderer.java
@@ -89,7 +89,7 @@ public class MergableGridRenderer implements IMergableGridRenderer {
         double x = 0;
         for ( int i = startColumnIndex; i <= endColumnIndex; i++ ) {
             final MergableGridColumn column = columns.get( i );
-            final int w = column.getWidth();
+            final double w = column.getWidth();
             if ( column.isLinked() ) {
                 final Rectangle lr = new Rectangle( w,
                                                     HEADER_HEIGHT )
@@ -131,7 +131,7 @@ public class MergableGridRenderer implements IMergableGridRenderer {
         for ( int i = startColumnIndex; i <= endColumnIndex; i++ ) {
             final MergableGridColumn column = columns.get( i );
             final Group hc = column.renderHeader();
-            final int w = column.getWidth();
+            final double w = column.getWidth();
             hc.setX( x + w / 2 )
                     .setY( getHeaderHeight() / 2 )
                     .setListening( false );
